[JAVA] Use a long to store the offset of a article in the zim file.
Fixes kiwix/kiwix-android#1769

+++ src/wrapper/java/org/kiwix/kiwixlib/Pair.java
@@ -22,5 +22,5 @@ package org.kiwix.kiwixlib;
 public class Pair
 {
   public String filename;
-  public int offset;
+  public long offset;
 }
